Added an option to recomposite Mip0 overtop of the result of the StackMipLevels method.
I didn't dig TOO deep into why the issue was happening but I found that the algorithm used in StackMipLevels was resulting in an image closer in quality to Mip1 than the original image. I added a "bool reCompositeMip0OnTop = true" arg to the method and if true will composite the original color onto the final image.